+++ pyriemann/embedding.py
@@ -7,16 +7,18 @@ from scipy.linalg import solve, eigh
 from scipy.sparse import csr_matrix
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.manifold import spectral_embedding
+from sklearn.manifold._utils import _binary_search_perplexity
 
 from .utils.distance import pairwise_distance
 from .utils.kernel import kernel as kernel_fct
+from .optimization.positive_definite import _get_tsne_embedding
 
 
 class SpectralEmbedding(BaseEstimator):
-    """Spectral embedding of SPD/HPD matrices into an Euclidean space.
+    """Spectral Embedding of SPD/HPD matrices.
 
-    It uses Laplacian Eigenmaps [1]_ to embed SPD/HPD matrices into an
-    Euclidean space of smaller dimension.
+    Spectral embedding uses Laplacian Eigenmaps [1]_ to embed SPD/HPD matrices
+    into an Euclidean space of smaller dimension.
     The basic hypothesis is that high-dimensional
     data live in a low-dimensional manifold, whose intrinsic geometry can be
     described via the Laplacian matrix of a graph. The vertices of this graph
@@ -28,7 +30,7 @@ class SpectralEmbedding(BaseEstimator):
     n_components : integer, default=2
         The dimension of the projected subspace.
     metric : string, default="riemann"
-        Metric used for defining pairwise distance between SPD matrices.
+        Metric used for defining pairwise distance between SPD/HPD matrices.
         For the list of supported metrics,
         see :func:`pyriemann.utils.distance.pairwise_distance`.
     eps : None | float, default=None
@@ -67,7 +69,7 @@ class SpectralEmbedding(BaseEstimator):
             eps = self.eps
 
         # make kernel matrix from the distance matrix
-        kernel = np.exp(-distmatrix ** 2 / (4 * eps))
+        kernel = np.exp(-(distmatrix**2) / (4 * eps))
 
         # normalize the kernel matrix
         q = kernel @ np.ones(len(kernel))
@@ -119,18 +121,18 @@ class SpectralEmbedding(BaseEstimator):
         -------
         X_new : ndarray, shape (n_matrices, n_components)
             Coordinates of embedded matrices.
-
         """
         self.fit(X)
         return self.embedding_
 
 
 class LocallyLinearEmbedding(TransformerMixin, BaseEstimator):
-    """Locally Linear Embedding (LLE) of SPD matrices.
+    """Locally Linear Embedding of SPD matrices.
 
-    As proposed in [1]_, Locally Linear Embedding (LLE) is a non-linear,
+    Locally Linear Embedding (LLE) is a non-linear,
     neighborhood-preserving dimensionality reduction algorithm which
-    consists of three main steps. For each SPD matrix X[i] [2]_:
+    consists of three main steps [1]_.
+    For each SPD matrix X[i] [2]_:
 
     1.  find its k-nearest neighbors k-NN(X[i]),
     2.  calculate the best reconstruction of X[i] based on its k-NN,
@@ -249,9 +251,10 @@ class LocallyLinearEmbedding(TransformerMixin, BaseEstimator):
         """
         _check_dimensions(self.data_, X)
         pairwise_dists = pairwise_distance(X, self.data_, metric=self.metric)
-        ind = np.array([
-            np.argsort(dist)[1:self.n_neighbors + 1] for dist in pairwise_dists
-        ])
+        ind = np.array(
+            [np.argsort(dist)[1: self.n_neighbors + 1]
+             for dist in pairwise_dists]
+        )
 
         weights = barycenter_weights(
             X,
@@ -286,8 +289,117 @@ class LocallyLinearEmbedding(TransformerMixin, BaseEstimator):
         return self.embedding_
 
 
+class TSNE(BaseEstimator):
+    """T-distributed Stochastic Neighbor Embedding (t-SNE) of SPD/HPD matrices.
+
+    T-distributed Stochastic Neighbor Embedding (t-SNE) reduces
+    a set of high-dimensional SPD/HPD matrices into
+    a set of low-dimensional SPD/HPD matrices [1]_.
+
+    Parameters
+    ----------
+    n_components : int, default=2
+        Low dimension of the matrices in the embedded space.
+    perplexity : int, default=None
+        Perplexity used in the t-SNE algorithm.
+        If None, it will be set to 0.75*n_matrices.
+    metric : {"euclid", "logeuclid", "riemann"}, default="riemann"
+        Metric for the gradient descent.
+    max_iter : int, default=200
+        Maximum number of iterations used for the gradient descent.
+    random_state : int, default=None
+        Pass an int for reproducible output across multiple function calls.
+
+    Attributes
+    ----------
+    embedding_ : ndarray, shape (n_matrices, n_components, n_components)
+        Embedding matrices of the training set.
+
+    Notes
+    -----
+    .. versionadded:: 0.9
+
+    References
+    ----------
+    .. [1] `Geometry-Aware visualization of high dimensional Symmetric
+        Positive Definite matrices
+        <https://openreview.net/pdf?id=DYCSRf3vby>`_
+        T. de Surrel, S. Chevallier, F. Lotte and F. Yger.
+        Transactions on Machine Learning Research, 2025
+    """
+
+    def __init__(
+        self,
+        n_components=2,
+        perplexity=None,
+        metric="riemann",
+        max_iter=200,
+        random_state=None,
+    ):
+        self.n_components = n_components
+        self.perplexity = perplexity
+        self.metric = metric
+        self.max_iter = max_iter
+        self.random_state = random_state
+
+    def fit(self, X, y=None):
+        """Fit TSNE.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of high-dimensional SPD/HPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        self : TSNE instance
+            The TSNE instance.
+        """
+
+        n_matrices, _, _ = X.shape
+        if self.perplexity is None:
+            self.perplexity = int(0.75 * n_matrices)
+
+        # Compute similarities between high-dimensional matrices
+        P = _compute_condprob_gaussian(X, self.metric, self.perplexity)
+
+        self.embedding_ = _get_tsne_embedding(
+            P,
+            self.n_components,
+            self.metric,
+            self.max_iter,
+            self.random_state,
+            _compute_jointprob_student
+        )
+
+        return self
+
+    def fit_transform(self, X, y=None):
+        """Calculate the embedded matrices.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of high-dimensional SPD/HPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        X_new : ndarray, shape (n_matrices, n_components, n_components)
+            Set of low-dimensional embedded matrices.
+        """
+        self.fit(X)
+        return self.embedding_
+
+
+###############################################################################
+
+
 def barycenter_weights(X, Y, indices, metric="riemann", kernel=None, reg=1e-3):
-    """Compute Riemannian barycenter weights of X from Y along the first axis.
+    """Compute barycenter weights of X from Y along the first axis.
 
     Estimates the weights to assign to each matrix in Y[indices] to recover
     the matrix X[i] by geodesic interpolation. The barycenter weights sum to 1.
@@ -321,8 +433,10 @@ def barycenter_weights(X, Y, indices, metric="riemann", kernel=None, reg=1e-3):
     .. versionadded:: 0.3
     """
     n_matrices, n_neighbors = indices.shape
-    msg = f"Number of index-sets in indices (is {n_matrices}) must match " \
-          f"number of matrices in X (is {X.shape[0]})."
+    msg = (
+        f"Number of index-sets in indices (is {n_matrices}) must match "
+        f"number of matrices in X (is {X.shape[0]})."
+    )
     assert X.shape[0] == n_matrices, msg
     if kernel is None:
         kernel = kernel_fct
@@ -407,7 +521,7 @@ def locally_linear_embedding(
     n_matrices, n_channels, n_channels = X.shape
     pairwise_distances = pairwise_distance(X, metric=metric)
     neighbors = np.array(
-        [np.argsort(dist)[1:n_neighbors + 1] for dist in pairwise_distances]
+        [np.argsort(dist)[1: n_neighbors + 1] for dist in pairwise_distances]
     )
 
     B = barycenter_weights(
@@ -438,28 +552,93 @@ def locally_linear_embedding(
     return embd, error
 
 
+def _compute_condprob_gaussian(X, metric, perplexity):
+    r"""Conditional probabilities using a Gaussian distribution.
+
+    ..math::
+        p_{j|i} = \frac{\exp(-\delta(X_i, X_j)^2/2\sigma_i^2)}{\
+        \sum_{k\neq i}\exp(-\delta(X_i, X_k)^2/2\sigma_i^2)}
+
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_matrices, n_channels, n_channels)
+        Set of SPD matrices.
+
+    Returns
+    -------
+    P : ndarray, shape (n_matrices, n_matrices)
+        Conditional probabilities of matrices.
+    """
+    n_matrices, _, _ = X.shape
+    Dsq = pairwise_distance(X, metric=metric, squared=True)
+    Dsq = Dsq.astype(np.float32, copy=False)
+    # Use _binary_search_perplexity from sklearn to compute conditional
+    # probabilities such that they approximately match the desired perplexity
+    conditional_P = _binary_search_perplexity(Dsq, perplexity, 0)
+
+    return conditional_P / n_matrices
+
+
+def _compute_jointprob_student(X, metric):
+    r"""Joint probabilities using a Student t-distribution with one DoF.
+
+    .. math::
+        p_{i,j} = \frac{\left(1 + \delta(X_i, X_j)^2\right)^{-1}}
+        {\sum_{k \neq l} (1 + \delta(X_k, X_l)^2)^{-1}}
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_matrices, n_channels, n_channels)
+        Set of SPD matrices.
+
+    Returns
+    -------
+    P : ndarray, shape (n_matrices, n_matrices)
+        Joint probabilities of matrices.
+    Dsq : ndarray, shape (n_matrices, n_matrices)
+        Squared distances between matrices.
+    """
+    n_matrices, _, _ = X.shape
+    Dsq = pairwise_distance(X, metric=metric, squared=True)
+
+    denominator = np.sum(
+        [np.sum([np.delete(1 / (1 + Dsq[k, :]), k)])
+         for k in range(n_matrices)]
+    )
+    P = 1 / (1 + Dsq) / denominator
+    np.fill_diagonal(P, 0)
+    return P, Dsq
+
+
 def _check_dimensions(X, Y=None, n_components=None, n_neighbors=None):
     n_matrices, n_channels, n_channels = X.shape
 
     if Y is not None and Y.shape[1:] != (n_channels, n_channels):
-        msg = f"Dimension of matrices in data to be transformed must match " \
-              f"dimension of data used for fitting. Expected " \
-              f"{(n_channels, n_channels)}, got {Y.shape[1:]}."
+        msg = (
+            f"Dimension of matrices in data to be transformed must match "
+            f"dimension of data used for fitting. Expected "
+            f"{(n_channels, n_channels)}, got {Y.shape[1:]}."
+        )
         raise ValueError(msg)
 
     if n_components is None:
         n_components = n_matrices - 1
     elif n_components >= n_matrices:
-        msg = f"n_components (is {n_components}) must be smaller than " \
-              f"n_matrices (is {n_matrices})."
+        msg = (
+            f"n_components (is {n_components}) must be smaller than "
+            f"n_matrices (is {n_matrices})."
+        )
         raise ValueError(msg)
 
     if n_neighbors is None:
         n_neighbors = n_matrices - 1
     elif n_matrices <= n_neighbors:
-        warnings.warn(f"n_neighbors (is {n_neighbors}) must be smaller than "
-                      f"n_matrices (is {n_matrices}). Setting n_neighbors to "
-                      f"{n_matrices - 1}.")
+        warnings.warn(
+            f"n_neighbors (is {n_neighbors}) must be smaller than "
+            f"n_matrices (is {n_matrices}). Setting n_neighbors to "
+            f"{n_matrices - 1}."
+        )
         n_neighbors = n_matrices - 1
 
     return n_components, n_neighbors