+++ doc/whatsnew.rst
@@ -12,6 +12,9 @@ v0.9.dev
 
 - Add conjugate transpose operator :func:`pyriemann.utils.base.ctranspose` for real- and complex-valued ndarrays. :pr:`348` by :user:`qbarthelemy`
 
+- Add :class:`pyriemann.embedding.TSNE`, a Riemannian t-SNE implementation
+  and update example comparing embeddings. :pr:`347` by :user:`thibaultdesurrel`
+
 v0.8 (February 2025)
 --------------------